+++ plugins/composeStrategy.go
@@ -47,11 +47,11 @@ func (s *composeStrategy) PruneExistingOffers(buyingAOffers []horizon.Offer, sel
 }
 
 // PreUpdate impl
-func (s *composeStrategy) PreUpdate(maxAssetBase float64, maxAssetQuote float64, trustBase float64, trustQuote float64) error {
+func (s *composeStrategy) PreUpdate(maxAssetBase float64, maxAssetQuote float64, trustBase float64, trustQuote float64, buyingAOffers []horizon.Offer, sellingAOffers []horizon.Offer) error {
 	// swap assets (base/quote) for buying strategy
-	e1 := s.buyStrat.PreUpdate(maxAssetQuote, maxAssetBase, trustQuote, trustBase)
+	e1 := s.buyStrat.PreUpdate(maxAssetQuote, maxAssetBase, trustQuote, trustBase, buyingAOffers, sellingAOffers)
 	// assets maintain same ordering for selling
-	e2 := s.sellStrat.PreUpdate(maxAssetBase, maxAssetQuote, trustBase, trustQuote)
+	e2 := s.sellStrat.PreUpdate(maxAssetBase, maxAssetQuote, trustBase, trustQuote, buyingAOffers, sellingAOffers)
 
 	if e1 == nil && e2 == nil {
 		return nil